Add multi-image handling
The code now will analyse a set of images individually and then
compile the results into one master bin list.

The way the gsqr bins and phi bins were defined had to be changed
to accommodate this. The bins are initialised before the images are
looped over so that they are consistent.