Roll back to version that worked on experiment
The code in its current satte has begun to stop working;
in particular the theta-phi images are not working. It has been
decided that the best option is to roll back to when this all
worked and re-add the functions from the broken code. This will
allow for the problem to be identified. The code will be tested
on xrd images after each addition to make sure the code isn't
broken.