Add polarisation correction and fix integration
The polarisation correction has been added, which corrects for the
polarisation caused by an incoming x-ray beam being scattered at
some angle. The integration has also been fixed such that the only
pixels that are binned are pixels within the user-set limits, and
pixels who have a value greater than 0. If a pixel has a value of
zero, or lies outside the user-defined range, the pixel is not
binned, but is counted as a discarded pixel.

Still to-do:
- multi-filter handling
- multi-image handling
- save/load data from previous run to save time in development,
and to allow reset points