Fixed G^2 calculation
There was a problem with the calculation of G^2 that caused the
symmetry of the co-ordinate system to reverse. This has now been
fixed. There was also a problem with getting the G^2 values to
match the expected values. This was because the wavelength was
simply set to the wrong value. This has been fixed now.

Still to-do:
- fix phi map
- add a depth parameter to sample attenuation
- make a compiled theta-phi image
- add multiple filter attenuation