Fixed compiled integrated intensity bug
There was a bug where integrated intensity would not be calculated
correctly when compiling the integrated intensities from multiple
images. This was because the intensities would be added together
without dividing by the number of images contributing to each
integrated intensity point.

This has been fixed by counting the number of images contributing
to each point in the compiled integrated_intensity_vs_gsqr image.
The summed intensity from the multiple images is now divided by the
number of contributing images.