Fixed calculations of phi and added phi0
It turns out there was a problem with how phi was being calculated.
Now, the vector connecting the origin and current pixel is collapsed
onto the phi0 plane, and the angle between this collapsed vector
and the phi0 vector is calculated. The phi0 vector is now defined
globally so that all images are on the same basis. This mean the
unwrapped images should be compilable.